+++ CotEditor/Sources/LineNumberView.swift
@@ -378,7 +378,6 @@ final class LineNumberView: NSRulerView {
         }()
         if ruleThickness != self.ruleThickness {
             self.ruleThickness = ruleThickness
-            self.textView?.invalidateTextContainerOrigin()
         }
     }